Add dummy index file; add rulesDirectory to tslint.json
The README's recommended way to use vrsource-tslint-rules's rules as
part of a larger tslint.json config is to add a `rulesDirectory` entry
with a relative path. However, this doesn't work if a `tslint.json`
has an indirect dependency on `vrsource-tslint-rules`.

For example, say my company has an `@mycompany/tslint-rules` package
whose `main` file is a `tslint.json`. If `@mycompany/tslint-rules` wants
to use a custom rule from `vrsource-tslint-rules`, it can try to add
`"rulesDirectory": "node_modules/vrsource-tslint-rules/rules"`. However,
although this path will be correct when `npm install` is run inside the
directory for `@mycompany/tslin-rules`, it will be wrong when a separate
`my-project` repo installs/depends on `@mycompany/tslint-rules`.

In that case, the tree will be:

```
my-project
- package.json
- node_modules
- @mycompany
- tslint-rules
- tslint.json
- vrsource-tslint-rules
- rules
```

So, the `tslint.json` file in `@mycompany/tslint-rules` would need to
reference `"../../vrsource-tslint-rules/rules"` as the `rulesDirectory`.

This all feels brittle (e.g., if `@mycompany/tslint-rules` was initially
an unscoped package that got moved to a scope, the paths would break),
and it's unnecessary. As @remyrylan notes in #13, you can (backwards-
compatibly) add `"rulesDirectory"` to `vrsource-tslint-rules`'s
`tslint.json`, and then consumers can import the custom rules with
`"extends"`.

Note: doing an `"extends"` does bring in vrsource's configuration for
all the built-in tslint rules, in addition to making the custom rules
available. This normally shouldn't be a problem, as consumers can easily
override the configuration they don't want (including by putting the
`"vrsource-tslint-rules"` earlier in the `"extends"` array than a base
`"tslint"` config they're using, like `"tslint:recommended").

However, where this is a problem, we can also make it possible to truly
just import the custom rules. To do that, we can add an `index.js`/
`index.ts` file to the rules directory; as the TSLint docs point out
(https://palantir.github.io/tslint/usage/configuration/), having an
index file leads TSLint to use node's module resolution algorithm on
the `"rulesDirectory"`. This means that consumers at any level can
specify `"rulesDirectory": "vrsource-tslint-rules/rules"` and have it
be resolved appropriately. This commit adds that index file too, and
updates the readme accordingly.